+++ docs/components/icons/CheckCircle.tsx
