+++ docs/components/icons/DismissIcon.tsx
