+++ docs/components/icons/ExternalLink.tsx
@@ -1,0 +1,21 @@
+import * as React from 'react';
+
+import * as Constants from '~/constants/theme';
+
+export const ExternalLink = ({ size = 24, stroke = Constants.expoColors.white }) => (
+  <svg
+    aria-label="check"
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6m4-3h6v6m-11 5L21 3"
+      stroke={stroke}
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+);
