+++ docs/components/icons/Info.tsx
@@ -1,3 +1,5 @@
+import * as React from 'react';
+
 import * as Constants from '~/constants/theme';
 
 export const Info = ({ size = 24, fill = Constants.expoColors.gray[600] }) => (
