+++ docs/components/icons/XCircle.tsx
@@ -1,17 +1,20 @@
+import * as React from 'react';
+
 import * as Constants from '~/constants/theme';
 
-export const ExternalLink = ({ size = 24, stroke = Constants.expoColors.white }) => (
+export const XCircle = ({ size = 24 }) => (
   <svg
-    aria-label="check"
+    aria-label="x"
     width={size}
     height={size}
     viewBox="0 0 24 24"
     fill="none"
     xmlns="http://www.w3.org/2000/svg">
+    <circle cx="12" cy="12" r="12" fill={Constants.expoColors.red[600]} />
     <path
-      d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6m4-3h6v6m-11 5L21 3"
-      stroke={stroke}
-      strokeWidth="2"
+      d="M16.5 7.5l-9 9M7.5 7.5l9 9"
+      stroke="#fff"
+      strokeWidth="1.5"
       strokeLinecap="round"
       strokeLinejoin="round"
     />
